\name{DatABEL-package}
\alias{DatABEL-package}
\title{DatABEL package...}
\description{DatABEL package}
\details{A package interfacing FILEVECTOR C++ library 
for storage of and fast consecutive access to 
large data matrices in out-of-RAM disk mode 
with regulated cache size. Columns of matrix 
are accessible very quickly.}
\seealso{\code{\link{make_empty_fvf}},
\code{\link{matrix2databel_base_R}},
\code{\linkS4class{databel_base_R-class}},
\code{\linkS4class{databel_filtered_R-class}}}
\author{Yurii Aulchenko}
