\name{databel}
\alias{databel}
\title{initiates databel object}
\usage{
  databel(baseobject, cachesizeMb = 64, readonly = TRUE)
}
\arguments{
  \item{baseobject}{name of the file or
  \link{databel-class} object}

  \item{cachesizeMb}{cache size (amount of RAM) to be used}

  \item{readonly}{readonly flag}
}
\description{
  this is a simple wrapper for "new" function creating
  databel object
}
\author{
  Yurii Aulchenko
}

