\name{databel_base_R}
\alias{databel_base_R}
\title{initiates databel_base_R object...}
\usage{databel_base_R(filename, cachesizeMb=64)}
\description{initiates databel_base_R object}
\details{initiates databel_base_R object}
\author{Yurii Aulchenko}
\arguments{\item{filename}{name of the file containing the data}
\item{cachesizeMb}{what cache size (size of RAM) to be used}}
