\name{databel_filtered_R}
\alias{databel_filtered_R}
\title{initiates databel_filtered_R object...}
\usage{databel_filtered_R(file_or_baseobject, cachesizeMb)}
\description{initiates databel_filtered_R object}
\details{initiates databel_filtered_R object}
\author{Yurii Aulchenko}
\arguments{\item{file_or_baseobject}{name of the file or \link{databel_base_R-class} object}
\item{cachesizeMb}{cache size (amount of RAM) to be used}}
