% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_text_file_columns}
\alias{extract_text_file_columns}
\title{extracts columns from text file}
\usage{
extract_text_file_columns(file, whichcols)
}
\arguments{
\item{file}{file name}

\item{whichcols}{which columns to extract}
}
\value{
matrix of strings with values from that columns
}
\description{
Extracts a column from text file to a matrix.
If in a particular file line the number of columns is less
then a column specified, returns last column!
}

