\name{DatABEL-package}
\alias{DatABEL-package}
\title{DatABEL package for fast consecutive access to large out-of-RAM stored matrices...}
\description{DatABEL package for fast consecutive access to large out-of-RAM stored matrices}
\details{A package interfacing FILEVECTOR C++ library 
for storage of and fast consecutive access to 
large data matrices in out-of-RAM disk mode 
with regulated cache size. Columns of matrix 
are accessible very quickly.}
\seealso{\code{\link{make_empty_fvf}},
\code{\link{matrix2databel}},
\code{\linkS4class{databel-class}}}
\author{Yurii Aulchenko (R code), Stepan Yakovenko & Andrey Chernyh (C++ code)}
