% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/databel.R
\name{databel}
\alias{databel}
\title{initiates databel object}
\usage{
databel(baseobject, cachesizeMb = 64, readonly = TRUE)
}
\arguments{
\item{baseobject}{name of the file or \link{databel-class} object}

\item{cachesizeMb}{cache size (amount of RAM) to be used}

\item{readonly}{readonly flag}
}
\description{
this is a simple wrapper for the "new" function
creating a databel object
}
\author{
Yurii Aulchenko
}

