% Generated by roxygen2 (4.0.1): do not edit by hand
\name{databel2text}
\alias{databel2text}
\title{Exports DatABEL object to a text file}
\usage{
databel2text(databel, file, NAString = "NA", row.names = TRUE,
  col.names = TRUE, transpose = FALSE)
}
\arguments{
\item{databel}{DatABEL object}

\item{file}{output file name}

\item{NAString}{string to replace NA with}

\item{row.names}{export row names if TRUE}

\item{col.names}{export col names if TRUE}

\item{transpose}{whether the matrix should be transposed}
}
\description{
Exports DatABEL object to a text file
}
\author{
Stepan Yakovenko
}

