\name{get_temporary_file_name}
\alias{get_temporary_file_name}
\title{generates temporary file name...}
\usage{get_temporary_file_name(path=".", withFVext=TRUE)}
\description{generates temporary file name}
\details{function to generate temporary file name}
\arguments{\item{path}{path to directory where
the temporary file will be located}
\item{withFVext}{whether function should check 
presence of *FVD and *FVI files too}}
