% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_temporary_file_name}
\alias{get_temporary_file_name}
\title{generates temporary file name}
\usage{
get_temporary_file_name(path = ".", withFVext = TRUE)
}
\arguments{
\item{path}{path to directory where the temporary file will be
located}

\item{withFVext}{whether function should check presence of *FVD and
*FVI files too}
}
\description{
function to generate temporary file name
}

